requirements.txt
aiohttp
websockets
python-dotenv
ccxt
cryptography
scipy
numpy
aiosqlite
